+++ backend/server/server.js
@@ -6,9 +6,12 @@ import path from 'path';
 import { fileURLToPath } from 'url';
 
 import connectDB from './config/db.js';
+import startScheduler from './utils/scheduler.js';
 
 import authRoutes from './routes/authRoutes.js';
 import inviteRoutes from './routes/inviteRoutes.js';
+import faqRoutes from './routes/faqRoutes.js';
+import postRoutes from './routes/postRoutes.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -16,6 +19,7 @@ const __dirname = path.dirname(__filename);
 const app = express();
 
 connectDB();
+startScheduler();
 
 app.use(cors({
     origin: process.env.FRONTEND_URL,
@@ -28,6 +32,8 @@ app.use(express.static(path.join(__dirname, '../../frontend')));
 
 app.use('/api/auth', authRoutes);
 app.use('/api/invite', inviteRoutes);
+app.use('/api/faq', faqRoutes);
+app.use('/api/posts', postRoutes);
 
 app.get('/', (req, res) => {
     res.send('Server is running!');